+++ tomcat/src/main/java/nextstep/jwp/controller/InternalServerExceptionController.java
@@ -16,7 +16,6 @@ public class InternalServerExceptionController implements org.apache.coyote.Exce
     @Override
     public void handle(Exception exception, Response response) throws IOException, URISyntaxException {
         log.error(exception.getMessage(), exception);
-        response.addHeader("Location", "/500.html");
-        response.writeWithEmptyBody(HttpStatus.FOUND);
+        response.redirect(HttpStatus.FOUND, "/500.html");
     }
 }
