+++ tomcat/src/main/java/nextstep/jwp/controller/requestmapping/ExceptionControllerFactory.java
@@ -1,6 +1,8 @@
-package nextstep.jwp.controller;
+package nextstep.jwp.controller.requestmapping;
 
 import java.util.List;
+import nextstep.jwp.controller.InternalServerExceptionController;
+import nextstep.jwp.controller.NotFoundExceptionController;
 import org.apache.coyote.ExceptionController;
 
 public class ExceptionControllerFactory {
