+++ tomcat/src/test/java/nextstep/org/apache/coyote/http11/Http11ProcessorTest.java
@@ -10,8 +10,8 @@ import java.io.IOException;
 import java.net.URL;
 import java.nio.file.Files;
 import org.apache.catalina.ControllerContainer;
-import nextstep.jwp.controller.ExceptionControllerFactory;
-import nextstep.jwp.controller.JwpRequestMapping;
+import nextstep.jwp.controller.requestmapping.ExceptionControllerFactory;
+import nextstep.jwp.controller.requestmapping.JwpRequestMapping;
 import org.apache.coyote.Container;
 import org.apache.coyote.http11.Http11Processor;
 import org.apache.coyote.http11.request.HttpMethod;
